Added a ClientHandler class to handle actions concerning the client and streams. Added a BookManager class to handle actions concerning books and the book list.
Testing with SocketTest yields good results when testing Get and GetAll + the help-command.

There are problems with the Add-method. No bugs occur, but the method doesn't work as intended. Fixing needed.

Delete and Update still await testing.

Additional Console.WriteLine() may improve the experience of viewing the server. Add this, if time allows it.